Add gitignore and clang-format for future repository changes
The plan is to switch this repository into a decompilation
project which implements Win32 code that is close to the original
assembly and compiles into an executable that can substitute for
the original.

After completion, a different project will be created as an
full engine reimplementation with portability and modern C
libraries in mind.

Markdown documentation is not sufficient on its own,
and there is no reason to place C definitions in Markdown when
they could be placed in real C files.

The clang-format file attempts to follow the EDK2 C Coding
Standards with some exceptions - most importantly, clang-format
cannot currently handle EDK2 function definitions.