Rename "win32/renderer" to "win32/graphics"
While reverse engineering the graphics module, I found that the
RestoreSurfaces() function actually makes a call to a large function
which appears to reload all textures.

This loading function is in a completely separate module in the
application code - it calls wrapper functions which use the global
instance of TmRenderer instantiated by main(), and renders
multiple frames to the screen on its own.

From looking further at the code, it appears that the current
"renderer" code that has been reverse engineering thus far is more
like a D3D/DirectDraw wrapper library, and that the real high-level
"renderer", which performs actions like rendering text and textures,
is more coupled to the application code.

To make this obvious, I've renamed the "renderer" code to "graphics".

Eventually, one of two things will have to happen:

 - "win32/graphics" will need to be a part of the "core/renderer"
   code and not its own standalone library
 - The texture reloading function will need to be exported to
   "win32/graphics" somehow