core: Add global C wrapper interface and overhaul symbol names
Evidence in the binary increasingly points to the idea that
SingleTrac maintained a public C API for the core engine code,
then changed the implementation of the engine for the Win32 PC
release of TM2.

This was not clear previously due to inlining and dead code
optimizations, and throws our current symbol naming scheme
into jeopardy.

The following changes have been made to help future reversing:

 - core/ reorganized
   - core/include houses the global include which all TM2 files
     should include
   - core/(operating system) holds the implementation of the TM2
     engine for that operating system
 - Symbols reorganized
   - Only public, exported TM2 engine C API functions should be
     prefixed with "Tm"
     - These functions should use normal C types, not Win32
     - Some of these functions will not exist in the actual binary
   - Module-level context objects (TmGraphics, TmSound, ...) have
     been renamed accordingly (GraphicsContext, SoundContext, etc.)
     - Global instances of these objects should be located next to
       their public C API

Additional changes in this commit:

 - Added doc/registry.md to document registry keys
 - Moved TmEnvironment to app/ and renamed to Environment
   - Rationale: There is no global variable for the Environment.
     This points to it being an application library.